Big rebuild of the front, and back ends. Standard CRUD ops work though
The front end now uses the look and feel from [1](Tuts+), and the back end
is mostly the https://github.com/gimler/symfony-rest-edition - but the API
doesn't appear to be entirely, clean, so I've had to hack the controllers
and entities to make it work. It particularly didn't like a full collection
with more than the expenses listed in the query() call (get all). Also
trouble with dates being posted back, so I've simplified that as well.

[1] http://code.tutsplus.com/courses/building-a-web-app-from-scratch-with-angularjs