big job, to add users to the system.
To avoind havingto complete fill the queries with WHERE clauses
I add a Doctrine Filter to (mostly) do it for me.
Still need to do it for the more complex raw-SQL queries though.
It's a lot easier than it was when I converted the GetUserID (which confirms
from userapp.io) into a service to inject as required.